EES-7102 Move ReleaseVersion and Publication interfaces to admin as no longer used on the frontend
