+++ src/explore-education-statistics-admin/src/pages/publication/__tests__/PublicationContactPage.test.tsx
@@ -488,7 +488,7 @@ describe('PublicationContactPage', () => {
       testContact.contactName,
     );
     expect(screen.getByTestId('Contact telephone')).toHaveTextContent(
-      testContact.contactTelNo,
+      testContact.contactTelNo as string,
     );
   });
 });
