+++ src/explore-education-statistics-admin/src/pages/release/content/components/ReleasePreviewTableTool.tsx
@@ -13,8 +13,10 @@ import mapTableHeadersConfig from '@common/modules/table-tool/utils/mapTableHead
 import tableBuilderService from '@common/services/tableBuilderService';
 import ButtonText from '@common/components/ButtonText';
 import ReleasePreviewTableToolFinalStep from '@admin/pages/release/content/components/ReleasePreviewTableToolFinalStep';
-import { Publication } from '@admin/services/publicationService';
-import { Publication as ContentPublication } from '@common/services/publicationService';
+import {
+  ContentPublication,
+  Publication,
+} from '@admin/services/publicationService';
 import React, { useState } from 'react';
 
 interface Props {
